This commit was manufactured by cvs2git to create branch 'ncurses-5_2-devel'.

+++ src/lib/libcurses/tinfo/parse_entry.c
@@ -245,7 +245,7 @@ _nc_parse_entry
 			(base == entryp->tterm.term_names)
 			? "primary name"
 			: "alias",
-			ptr - base, base);
+			(int)(ptr - base), base);
 	}
     }
